One segment movement implemented & some food spawner methods implemented.

+++ core/src/com/vsu/csf/utils/Consts.java
@@ -1,5 +1,5 @@
 package com.vsu.csf.utils;
 
 public class Consts {
-    public static final double RADIANS_TO_DEGREES = 57.2958;
+    public static final float RADIANS_TO_DEGREES = 57.2958f;
 }
